fix uploadSlip
